# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "Jace NLP App"
  description: "API to retrieve simply NER entities on English (en), Russian (ru), Ukrainian (uk) languages."
  contact:
    email: "its.an.omen.man@gmail.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: localhost
basePath: /
#
tags:
- name: "NER"
  description: "Named-entity recognition (NER)"
#
schemes:
  - http
  - https
consumes:
  - text/plain
produces:
  - text/plain
paths:
  /ner/tokenize:
    post:
      tags:
      - "NER"
      summary: Return list of tokens from text
      description: "Split string into tokens and return them as a list"
      operationId: tokenize
      consumes:
      - text/plain
      parameters:
      - in: body
        name: text
        description: Input text
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Success"
          schema:
            type: string
        "405":
          description: "Invalid input"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /ner/tokenize_with_offsets:
    post:
      tags:
      - "NER"
      summary: Return tuple list of tokens and their positions from text
      description: "Split string into tokens and return them as a list.  Also, each element of the list
       contains a tuple of the token text and the byte offset which indicates the position of the
       first character in the token within the input string."
      operationId: tokenize_with_offsets
      consumes:
      - text/plain
      parameters:
      - in: body
        name: text
        description: Input text
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Success"
          schema:
            type: string
        "405":
          description: "Invalid input"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /ner/get_possible_ner_tags:
    post:
      tags:
      - "NER"
      summary: Return possible ner tags
      description: "Return possible ner tags for each lang model (en, ru, uk)"
      operationId: get_possible_ner_tags
      consumes:
      - text/plain
      parameters:
      - in: query
        name: lang
        description: Specify language (en, ru, uk)
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Success"
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /ner/extract_entities:
    post:
      tags:
      - "NER"
      summary: Return NER entities from text
      description: "Return NER entities from text"
      operationId: extract_entities
      consumes:
      - text/plain
      parameters:
      - in: body
        name: text
        description: Input text
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Success"
          schema:
            type: string
        "405":
          description: "Invalid input"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /ner/extract_entities_pretty:
    post:
      tags:
      - "NER"
      summary: Return NER entities from text
      description: "Return NER entities from text (pretty)"
      operationId: extract_entities_pretty
      consumes:
      - text/plain
      parameters:
      - in: body
        name: text
        description: Input text
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Success"
          schema:
            type: string
        "405":
          description: "Invalid input"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string